// apps/hermes-studio/src/types/gateway.ts
export interface TodoItem {
  id: string;
  content: string;
  status: 'cancelled' | 'completed' | 'in_progress' | 'pending';
}